#416 auto init of budibase when electron starts
